Looking up test function if given a symbol as argument
The empty-dict and empty-set functions used to take the `test` argument
and immediately store it as the equality predicate for the dict/set.
However, if you pass 'equal and #'equal, in one case you get a symbol
and in the other a function; this patch looks up the relevant function
in case a symbol is passed.